Corrected issues with charshift with custom encoding
current_char_ord value was not updating due to typo, corrected.
Discovered that value of get_shift_position could be equal to the collection_length,
this caused a nil value to be generated when gem attempted to get the character at that
position in the encoding (over by one.)  Added new conditional to return 0 if return
value of get_shift_position would be the encoding length.
Expanded test to include multi-char strings and overflow condition for custom encodings.